resolve changes by @ruvmello
